+++ pg-core/src/java/org/pg/util/SQL.java
@@ -11,7 +11,7 @@ public final class SQL {
     public static String SQLSetTxReadOnly = "SET TRANSACTION READ ONLY";
 
     public static String SQLSetTxLevel (final TxLevel level) {
-        return String.format("SET TRANSACTION ISOLATION LEVEL %s", level.getCode());
+        return "SET TRANSACTION ISOLATION LEVEL " + level.getCode();
     }
 
     public static void main (final String[] args) {